==> tasks/4. transform files/0. combine files.ipynb <==
# Combine files <span style="color:red">EXCLUDES!</span>
This script combines all files to one, big, wunderfull, shiney file. Do this after the manual check!

import glob
import ntpath

LIST_FOLDER = '../../data/2. manual check/lists/*.csv'
LIST_OUT = '../../data/3. transformation/0_list_2019.csv'

ACCUM_FILDER = '../../data/2. manual check/accumulations/*.csv'
ACCUM_OUT = '../../data/3. transformation/0_accumulations_2019.csv'

#EXCLUDE_LIST = ['menarini.csv', 'novartis.csv', 'recordati.csv', 'takeda.csv', 'pfizer.csv']
EXCLUDE_LIST = []

def combine_files(folder_in, file_out):
    print(EXCLUDE_LIST)
    #Read Folder
    csv_list = glob.glob(folder_in)
    csv_list.sort()
    
    #Loop files
    firstfile = True
    with open(file_out, 'w', encoding='UTF-8') as outfile:
        for fname in csv_list:
            if ntpath.basename(fname) in EXCLUDE_LIST:
                print("Excluded: %s" % fname)
                continue;
            
            with open(fname, encoding='UTF-8') as infile:
                #print(fname)
                i = 0
                for line in infile:
                    if firstfile or ( not firstfile and i > 0):
                        outfile.write(line)
                    i += 1
                #When not empty file, add \n
                if i > 1:
                    outfile.write('\n')
                firstfile = False
    print(str(len(csv_list)) + " Files combined in " + folder_in)


combine_files(LIST_FOLDER, LIST_OUT)
combine_files(ACCUM_FILDER, ACCUM_OUT)
