==> tasks/4. transform files/1. clean accumulations.ipynb <==
# 1. Clean Accumulations
Remove counts. Not reliable

import pandas as pd
import numpy as np
import sys
sys.path.insert(0, '../../data/lib/')
import consts
import re

## Import

df_raw = pd.read_csv('../../data/3. transformation/0_accumulations_2019.csv', encoding='UTF-8')
df_data = df_raw.copy()

## Remove count - NOT ANYMORE!

#df_data = df_data[df_data.type.isin(['hcp_amount', 'hco_amount', 'rnd'])]

## Save!

df_data.to_csv('../../data/3. transformation/2_accumulations_cleaned.csv', index=False)
