==> tasks/4. transform files/3. check dataset.ipynb <==
# 3. Check Dataset
This file performes some checks. They must not all workout, but you will have far bether results by fixing it!

import pandas as pd
import numpy as np
import sys

df_raw = pd.read_csv('../../data/3. transformation/2_list_all_cleaned.csv')
df_accu = pd.read_csv('../../data/3. transformation/2_accumulations_cleaned.csv')

## Empty HCP-Address check
You should not have any rows withouth an address

#df_raw[df_raw.address.isna() & (df_raw.type == 'hcp')]

## Check for empty name

df_raw[df_raw.name.isna()]

## Check Sources
Are sources correct?

df_sources = pd.read_csv('../6. database export/sources/liste_companies.csv')

#Join
df_raw = df_raw.merge(df_sources, how='left', left_on='source', right_on='pha_key')
df_raw[df_raw.pha_name.isna()]

df_sources = pd.read_csv('../6. database export/sources/liste_companies.csv')
    
#Join
df_accu = df_accu.merge(df_sources, how='left', left_on='source', right_on='pha_key')
df_accu[df_accu.pha_name.isna()]

## Check Schweiz. or Schw.
Replace them!

df_raw[df_raw.name.str.contains("Schweiz.", regex=False, case=False) | df_raw.name.str.contains("Schw.", regex=False, case=False)]

## Check wrong Country

df_raw[df_raw.country != 'CH']
