==> tasks/4. transform files/4. expand columns.ipynb <==
# 4. Expand Columns
This file expands the different columns.
If you dont like to install libpostal (its huge!), skip the address-step

from postal.expand import expand_address

import pandas as pd
import numpy as np
import sys
import re

import sys
sys.path.insert(0, '../../data/lib/')
from consts import *

## Import

df_new = pd.read_csv('../../data/3. transformation/2_list_all_cleaned.csv', dtype = dataframe_types)
df_live = pd.read_csv('../../data/0. live data/0_recipient.csv')

## Expand Address

def expand_df(df_data):
    
    #expand Addres
    df_data['address'] = df_data['address'].fillna("")
    df_data['location'] = df_data['location'].fillna("")

    df_data.loc[df_data.address != "", 'address_expand'] = df_data.loc[df_data.address != "", 'address'].apply(lambda x: expand_address(x)[0])
    df_data.loc[df_data.location != "", 'location_expand'] = df_data.loc[df_data.location != "", 'location'].apply(lambda x: expand_address(x)[0])
    
    #Expand Name
    df_data['name_expand'] = df_data['name']

    # Remove "Schweizerische Gesellschaft für"
    df_data['name_expand'] = df_data['name_expand'].str.replace('Schweizerische Gesellschaft für', '')
    df_data['name_expand'] = df_data['name_expand'].str.replace('schweizerische ges. für', '')

    # Remove (c/o)
    df_data['name_expand'] = df_data['name_expand'].str.replace(r'(?i)c\/o.*', '', regex=True)

    #Lowercase
    df_data['name_expand'] = df_data['name_expand'].str.lower()
    
    return df_data

## save

expand_df(df_new).to_csv('../../data/3. transformation/3_list_expanded.csv', index=False)
expand_df(df_live).to_csv('../../data/0. live data/1_recipient_expanded.csv', index=False)
