==> tasks/4. transform files/5. expand_addresses.ipynb <==
# 4. Expand addresses
This file expands the addresses with libpostal. We have done this in a separate step so we could upload the expanded file to our server and run the pair finding there. Otherwise you have to install the huge libpostal library on the server.

from postal.expand import expand_address

import pandas as pd
import numpy as np
import sys
import re

df = pd.read_csv('../../data/3. transformation/2_list_all_cleaned.csv')

df_data = df.copy()

df_data['address'] = df_data['address'].astype("str")
df_data['location'] = df_data['location'].astype("str")

df_data.loc[df_data.address.notna(), 'address_expand'] = df_data['address'].apply(lambda x: expand_address(x)[0])
df_data.loc[df_data.location.notna(), 'location_expand'] = df_data['location'].apply(lambda x: expand_address(x)[0])

df_data.to_csv('../../data/3. transformation/3_list_expanded.csv', index=False)
